file move didn't get checked
